feat : AuthAcceptanceTest
 - 테스트추가
