+++ src/test/java/wooteco/subway/acceptance/member/MemberAcceptanceTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import wooteco.subway.AcceptanceTest;
 import wooteco.subway.service.member.dto.MemberResponse;
+import wooteco.subway.service.member.dto.TokenResponse;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -15,13 +16,15 @@ public class MemberAcceptanceTest extends AcceptanceTest {
         String location = createMember(TEST_USER_EMAIL, TEST_USER_NAME, TEST_USER_PASSWORD);
         assertThat(location).isNotBlank();
 
-        MemberResponse memberResponse = getMember(TEST_USER_EMAIL);
+        TokenResponse token = login(TEST_USER_EMAIL, TEST_USER_PASSWORD);
+
+        MemberResponse memberResponse = getMember(token);
         assertThat(memberResponse.getId()).isNotNull();
         assertThat(memberResponse.getEmail()).isEqualTo(TEST_USER_EMAIL);
         assertThat(memberResponse.getName()).isEqualTo(TEST_USER_NAME);
 
         updateMember(memberResponse);
-        MemberResponse updatedMember = getMember(TEST_USER_EMAIL);
+        MemberResponse updatedMember = getMember(token);
         assertThat(updatedMember.getName()).isEqualTo("NEW_" + TEST_USER_NAME);
 
         deleteMember(memberResponse);
