+++ src/test/java/wooteco/subway/web/line/LineControllerTest.java
@@ -4,18 +4,22 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import wooteco.subway.config.ETagHeaderFilter;
+import wooteco.subway.config.WebMvcConfig;
 import wooteco.subway.web.LineController;
 import wooteco.subway.domain.line.Line;
 import wooteco.subway.domain.station.Station;
 import wooteco.subway.service.line.LineService;
 import wooteco.subway.service.line.dto.LineDetailResponse;
 import wooteco.subway.service.line.dto.WholeSubwayResponse;
+import wooteco.subway.web.member.interceptor.BearerAuthInterceptor;
 
 import java.util.Arrays;
 import java.util.List;
@@ -26,9 +30,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.header;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@Disabled
-@WebMvcTest(controllers = LineController.class)
-@Import(ETagHeaderFilter.class)
+@SpringBootTest
+@AutoConfigureMockMvc
+@Import({ETagHeaderFilter.class})
 public class LineControllerTest {
     @MockBean
     private LineService lineService;
